#625 refactor(common): Correct some mistakes
